Make the halo excange method user-tweakable
Now the user can set HIPBONE_OGS_METHOD to either

pairwise
alltoall
crystalrouter
auto

to force hipbone to use a specific communication method.

If the variable is not set, or is set to the empty string, then hipbone
will use 'auto' and time each at runtime and choose the fastest one.

If the variable is set to anything else, it is an error.